Add config.yml & loader, move callbacks to diff modules
